First cut at core.match-based dispatcher
io.aviso.rook.dispatcher/compile-dispatch-table compiles dispatch tables
to Clojure functions that dispatch directly to the handler function
wrapped in the specified middleware/pipeline.

A dispatch table is a vector of vectors of the following shape:

  [[method pathvec verb-fn pipeline
    [method' pathvec' verb-fn' pipeline' ...]
    ...]
   ...]

where a pathvec is a vector of strings representing literal path
segments and/or keywords representing path segments to be bound to route
parameters named by the keywords.