First cut at namespace-dispatch-table
Dispatch tables can now include "context entries" distinguished by the
omission of the method keyword. The pathvec of such entries serves as
the context-pathvec for the nested entries. Providing a pipeline is
optional; if present, it will be used as the default pipeline for the
nested entries.

Regular entries can now omit the pipeline field or specify nil to use
the default pipeline inherited from the context.