Implement significant revision to io.aviso.rook.dispatcher
Simplifies the construction of the handler as the remaining artifacts of the long-removed code generation support
are gone.  It also clarifies how child namespaces extend container namespaces.

Additionally, this adds the ability to define argument resolvers for a namespace, which applies to the namespace
and to its children (which can further override the argument resolvers).

This currently breaks the Swagger integration.